move AesMode and AesVendorVersion out of aes-crypto feature
