+++ src/read.rs
@@ -1,13 +1,13 @@
 //! Types for reading ZIP archives
 
 #[cfg(feature = "aes-crypto")]
-use crate::aes::{AesMode, AesReader, AesReaderValid, AesVendorVersion};
+use crate::aes::{AesReader, AesReaderValid};
 use crate::compression::CompressionMethod;
 use crate::cp437::FromCp437;
 use crate::crc32::Crc32Reader;
 use crate::result::{InvalidPassword, ZipError, ZipResult};
 use crate::spec;
-use crate::types::{DateTime, System, ZipFileData};
+use crate::types::{AesMode, AesVendorVersion, DateTime, System, ZipFileData};
 use crate::zipcrypto::{ZipCryptoReader, ZipCryptoReaderValid, ZipCryptoValidator};
 use byteorder::{LittleEndian, ReadBytesExt};
 use std::borrow::Cow;
@@ -175,7 +175,7 @@ fn make_crypto_reader<'a>(
     using_data_descriptor: bool,
     reader: io::Take<&'a mut dyn io::Read>,
     password: Option<&[u8]>,
-    #[cfg(feature = "aes-crypto")] aes_info: Option<(AesMode, AesVendorVersion)>,
+    aes_info: Option<(AesMode, AesVendorVersion)>,
     #[cfg(feature = "aes-crypto")] compressed_size: u64,
 ) -> ZipResult<Result<CryptoReader<'a>, InvalidPassword>> {
     #[allow(deprecated)]
@@ -185,10 +185,13 @@ fn make_crypto_reader<'a>(
         }
     }
 
-    #[cfg(not(feature = "aes-crypto"))]
-    let aes_info: Option<()> = None;
-
     let reader = match (password, aes_info) {
+        #[cfg(not(feature = "aes-crypto"))]
+        (Some(_), Some(_)) => {
+            return Err(ZipError::UnsupportedArchive(
+                "AES encrypted files cannot be decrypted without the aes-crypto feature.",
+            ))
+        }
         #[cfg(feature = "aes-crypto")]
         (Some(password), Some((aes_mode, vendor_version))) => {
             match AesReader::new(reader, aes_mode, compressed_size).validate(&password)? {
@@ -210,7 +213,8 @@ fn make_crypto_reader<'a>(
                 Some(r) => CryptoReader::ZipCrypto(r),
             }
         }
-        _ => CryptoReader::Plaintext(reader),
+        (None, Some(_)) => return Ok(Err(InvalidPassword)),
+        (None, None) => CryptoReader::Plaintext(reader),
     };
     Ok(Ok(reader))
 }
@@ -549,7 +553,6 @@ impl<R: Read + io::Seek> ZipArchive<R> {
             data.using_data_descriptor,
             limit_reader,
             password,
-            #[cfg(feature = "aes-crypto")]
             data.aes_mode,
             #[cfg(feature = "aes-crypto")]
             data.compressed_size,
@@ -646,7 +649,6 @@ pub(crate) fn central_header_to_zip_file<R: Read + io::Seek>(
         data_start: 0,
         external_attributes: external_file_attributes,
         large_file: false,
-        #[cfg(feature = "aes-crypto")]
         aes_mode: None,
     };
 
@@ -696,7 +698,6 @@ fn parse_extra_field(file: &mut ZipFileData) -> ZipResult<()> {
                     len_left -= 8;
                 }
             }
-            #[cfg(feature = "aes-crypto")]
             0x9901 => {
                 // AES
                 if len != 7 {
@@ -1048,7 +1049,6 @@ pub fn read_zipfile_from_stream<'a, R: io::Read>(
         // from standard input, this field is set to zero.'
         external_attributes: 0,
         large_file: false,
-        #[cfg(feature = "aes-crypto")]
         aes_mode: None,
     };
 
@@ -1075,7 +1075,6 @@ pub fn read_zipfile_from_stream<'a, R: io::Read>(
         result.using_data_descriptor,
         limit_reader,
         None,
-        #[cfg(feature = "aes-crypto")]
         None,
         #[cfg(feature = "aes-crypto")]
         result.compressed_size,
