+++ src/types.rs
@@ -1,8 +1,5 @@
 //! Types that specify what is contained in a ZIP.
 
-#[cfg(feature = "aes-crypto")]
-use crate::aes::{AesMode, AesVendorVersion};
-
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum System {
     Dos = 0,
@@ -251,7 +248,6 @@ pub struct ZipFileData {
     pub external_attributes: u32,
     /// Reserve local ZIP64 extra field
     pub large_file: bool,
-    #[cfg(feature = "aes-crypto")]
     /// AES mode if applicable
     pub aes_mode: Option<(AesMode, AesVendorVersion)>,
 }
@@ -301,6 +297,39 @@ impl ZipFileData {
     }
 }
 
+/// The encryption specification used to encrypt a file with AES.
+///
+/// According to the [specification](https://www.winzip.com/win/en/aes_info.html#winzip11) AE-2
+/// does not make use of the CRC check.
+#[derive(Copy, Clone, Debug)]
+pub enum AesVendorVersion {
+    Ae1,
+    Ae2,
+}
+
+/// AES variant used.
+#[derive(Copy, Clone, Debug)]
+pub enum AesMode {
+    Aes128,
+    Aes192,
+    Aes256,
+}
+
+#[cfg(feature = "aes-crypto")]
+impl AesMode {
+    pub fn salt_length(&self) -> usize {
+        self.key_length() / 2
+    }
+
+    pub fn key_length(&self) -> usize {
+        match self {
+            Self::Aes128 => 16,
+            Self::Aes192 => 24,
+            Self::Aes256 => 32,
+        }
+    }
+}
+
 #[cfg(test)]
 mod test {
     #[test]
@@ -335,7 +364,6 @@ mod test {
             central_header_start: 0,
             external_attributes: 0,
             large_file: false,
-            #[cfg(feature = "aes-crypto")]
             aes_mode: None,
         };
         assert_eq!(
